requirements.txt
# Core dependencies
nemo-guardrails==0.7.1
openai==1.3.7
groq==0.4.1

# PyTorch (adjust based on your CUDA version)
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Additional dependencies for NeMo Guardrails
transformers>=4.21.0
sentence-transformers>=2.2.0
langchain>=0.0.340
pydantic>=2.0.0
pyyaml>=6.0
aiohttp>=3.8.0
fastapi>=0.104.0
uvicorn>=0.24.0

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# Logging and monitoring
python-json-logger>=2.0.0
rich>=13.0.0

# Environment management
python-dotenv>=1.0.0